+++ sdk/appconfiguration/azure-appconfiguration/azure/appconfiguration/_generated/models/_models.py
@@ -6,7 +6,7 @@
 # --------------------------------------------------------------------------
 
 from msrest.serialization import Model
-from azure.core import HttpResponseError
+from azure.core.exceptions import HttpResponseError
 
 
 class Error(Model):
