+++ sdk/keyvault/azure-keyvault-certificates/azure/keyvault/certificates/_shared/_generated/v2016_10_01/models/_models.py
@@ -10,7 +10,7 @@
 # --------------------------------------------------------------------------
 
 from msrest.serialization import Model
-from azure.core import HttpResponseError
+from azure.core.exceptions import HttpResponseError
 
 
 class Action(Model):
