+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/download_async.py
@@ -9,7 +9,7 @@ import sys
 from io import BytesIO
 from itertools import islice
 
-from azure.core import HttpResponseError
+from azure.core.exceptions import HttpResponseError
 from .._shared.encryption import decrypt_blob
 from .._shared.request_handlers import validate_and_format_range_headers
 from .._shared.response_handlers import process_storage_error, parse_length_from_content_range
