+++ sdk/storage/azure-storage-blob/tests/test_block_blob_sync_copy_async.py
@@ -7,7 +7,7 @@ import pytest
 import asyncio
 
 from datetime import datetime, timedelta
-from azure.core import HttpResponseError
+from azure.core.exceptions import HttpResponseError
 from azure.core.pipeline.transport import AioHttpTransport
 from multidict import CIMultiDict, CIMultiDictProxy
