+++ sdk/storage/azure-storage-file/azure/storage/file/download.py
@@ -8,7 +8,7 @@ import sys
 import threading
 from io import BytesIO
 
-from azure.core import HttpResponseError
+from azure.core.exceptions import HttpResponseError
 from azure.core.tracing.common import with_current_context
 from ._shared.encryption import decrypt_blob
 from ._shared.request_handlers import validate_and_format_range_headers
